refactor: atom 모듈화
