+++ src/components/CartController/CartController.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent } from 'react';
 import * as S from './CartController.style';
-import * as T from '../../types/ProductType';
+import * as T from '../../types/types';
 import cartIcon from '../../assets/cart.svg';
 import useCart from '../../hooks/useCart';
